+++ back-end/mogul/src/main/java/com/mogul/demo/board/service/CommentServiceImpl.java
@@ -51,7 +51,7 @@ public class CommentServiceImpl implements CommentService{
 
 	@Override
 	@Transactional
-	public boolean removeComment(int id) {
+	public boolean removeComment(Long id) {
 		Optional<Comment> comment = Optional.ofNullable(
 			commentRepository.findById(id).orElseThrow(() -> new EntityNotFoundException("해당 댓글을 찾을 수 없습니다")));
 		comment.get().deleteComment();
